+++ academy-ui/src/app/api/[spaceId]/queries/spaces/by-creator/route.ts
@@ -10,13 +10,14 @@ export async function GET(req: NextRequest): Promise<NextResponse<Space[] | { er
     return NextResponse.json({ error: 'Creator username is required' }, { status: 400 });
   }
 
-  const spaceIds = await prisma.user
+  const spaceMappings = await prisma.user
     .findMany({
       where: { username: username },
       select: { spaceId: true },
     })
-    .then((users) => users.map((user) => user.spaceId));
-
+    
+  const spaceIds = spaceMappings.map(mapping => mapping.spaceId);
+  
   if (spaceIds.length === 0) {
     return NextResponse.json([], { status: 200 });
   } else {
