+++ academy-ui/src/app/api/[spaceId]/queries/users/by-username/route.ts
@@ -3,7 +3,7 @@ import { prisma } from '@/prisma';
 import { withErrorHandlingV1 } from '@/app/api/helpers/middlewares/withErrorHandling';
 import { UserDto } from '@/types/user/UserDto';
 
-async function getHandler(req: NextRequest): Promise<NextResponse<UserDto[] | { error: string }>> {
+async function getHandler(req: NextRequest): Promise<NextResponse<UserDto | { error: string }>> {
   const { searchParams } = new URL(req.url);
   const username = searchParams.get('username');
 
@@ -15,7 +15,7 @@ async function getHandler(req: NextRequest): Promise<NextResponse<UserDto[] | { 
     where: { username: username },
   });
 
-  return NextResponse.json([user] as UserDto[], { status: 200 });
+  return NextResponse.json(user as UserDto, { status: 200 });
 }
 
-export const GET = withErrorHandlingV1<UserDto[]>(getHandler);
+export const GET = withErrorHandlingV1<UserDto>(getHandler);
