+++ academy-ui/src/app/spaces/finish-space-setup/page.tsx
@@ -1,7 +1,7 @@
 import { getSpaceServerSide } from '@/utils/space/getSpaceServerSide';
 import FinishSpaceSetup from '@/components/spaces/finishSetup/FinishSpaceSetup';
 
-export default async function CreateSpacePage() {
+export default async function FinishSpaceSetupPage() {
   const space = await getSpaceServerSide();
   return <FinishSpaceSetup space={space!} />;
 }
